refactored files: styles folder in components
