+++ src/components/Header.js
@@ -1,4 +1,4 @@
-import '../styles/Header.css';
+import './styles/Header.css';
 import React from "react";
 import Logo from "./Logo";
 import SearchBar from "./SearchBar";
