+++ src/components/styles/LeftPanel.css
