+++ src/components/styles/RightPanel.css
