+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
-import './styles/Header.css';
+import './components/styles/Header.css';
 import App from "./components/App";
 
 ReactDOM.render(
